feat: enhance token limit management and error handling across AI components
- Updated `checkLimit` method to consider input and output tokens.
- Improved token calculation logic in AI request providers.
- Enhanced JSON parsing in `JsonUtil` with Optional for safer operations.
- Added detailed error handling in `OpenAIAPIInterceptor` for invalid requests.
- Refined token estimation logic with safety margins and JSON structure considerations.

+++ core/src/main/java/com/predic8/membrane/core/interceptor/ai/provider/ClaudeAiRequest.java
@@ -14,7 +14,7 @@ public class ClaudeAiRequest extends AbstractAiApiRequest {
     @Override
     public long estimateInputTokens() {
         // System prompt
-        int tokens = json.path("system").asText().length() / 4;
+        long tokens = json.path("system").asText().length() / 4;
 
         // Messages
         for (var message : json.path("messages")) {
